requirements.txt
absl-py==0.13.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
argcomplete==1.10.3
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.8.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
compressed-rtf==1.0.6
config==0.5.1
cryptography==36.0.1
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
Django==3.2.9
djangorestframework==3.12.4
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.28.7
flatbuffers==1.12
frozenlist==1.2.0
gast==0.4.0
google-api-core==1.31.5
google-api-python-client==1.12.8
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
Google-Images-Search==1.3.10
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
httplib2==0.20.2
idna==2.10
IMAPClient==2.1.0
importlib-metadata==4.8.2
ipykernel==6.2.0
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==7.0.1
jupyter-core==4.7.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.7.1
Markdown==3.3.4
matplotlib==3.4.1
matplotlib-inline==0.1.2
multidict==5.2.0
nest-asyncio==1.5.1
nltk==3.6.7
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.2.4
parso==0.8.2
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.12.0
pyee==8.2.2
pyfiglet==0.8.post1
Pygments==2.10.0
PyMuPDF==1.19.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppeteer==0.2.6
python-dateutil==2.8.1
python-pptx==0.6.21
python-resize-image==1.1.20
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==301
pyzbar==0.1.8
pyzmq==22.2.1
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.2
seaborn==0.11.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1+cu102
torchaudio==0.8.1
torchvision==0.9.1+cu102
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.1
windows-curses==2.3.0
wrapt==1.12.1
wsproto==1.0.0
xarray==0.17.0
xlrd==1.2.0
XlsxWriter==3.0.2
yarl==1.7.2
zipp==3.6.0
